Extract CourseGridCell
