+++ ArtemisKit/Sources/Dashboard/CourseGridCell.swift
@@ -1,63 +1,16 @@
 //
-//  CourseGridView.swift
+//  CourseGridCell.swift
 //
 //
-//  Created by Nityananda Zbil on 29.11.23.
+//  Created by Nityananda Zbil on 19.03.24.
 //
 
-import CourseRegistration
-import DesignLibrary
+import DesignLibrary //
 import Navigation
 import SharedModels
 import SwiftUI
 
-struct CourseGridView: View {
-
-    private static let layout = [GridItem(.adaptive(minimum: 400, maximum: .infinity), spacing: .l, alignment: .center)]
-
-    @ObservedObject var viewModel: DashboardViewModel
-    @State private var isCourseRegistrationPresented = false
-
-    var body: some View {
-        DataStateView(data: $viewModel.coursesForDashboard) {
-            await viewModel.loadCourses()
-        } content: { coursesForDashboard in
-            ScrollView {
-                LazyVGrid(columns: Self.layout, spacing: .l) {
-                    ForEach(coursesForDashboard.courses ?? [], content: CourseGridCellView.init(courseForDashboard:))
-                }
-                .padding(.horizontal, .l)
-
-                HStack {
-                    Spacer()
-                    Button(R.string.localizable.dashboardRegisterForCourseButton()) {
-                        isCourseRegistrationPresented = true
-                    }
-                    .buttonStyle(ArtemisButton())
-                    Spacer()
-                }
-            }
-            .refreshable {
-                await viewModel.loadCourses()
-            }
-        }
-        .sheet(isPresented: $isCourseRegistrationPresented) {
-            CourseRegistrationView {
-                isCourseRegistrationPresented = false
-                viewModel.coursesForDashboard = .loading
-                Task {
-                    await viewModel.loadCourses()
-                }
-            }
-        }
-        .task {
-            await viewModel.loadCourses()
-        }
-    }
-}
-
-private struct CourseGridCellView: View {
-
+struct CourseGridCell: View {
     @EnvironmentObject private var navigationController: NavigationController
 
     let courseForDashboard: CourseForDashboardDTO
@@ -89,7 +42,7 @@ private struct CourseGridCellView: View {
     }
 }
 
-private extension CourseGridCellView {
+private extension CourseGridCell {
     var header: some View {
         HStack(alignment: .center) {
             AsyncImage(url: courseForDashboard.course.courseIconURL) { phase in
@@ -169,8 +122,4 @@ private extension CourseGridCellView {
         .background(Color.Artemis.dashboardCardBackgroundColor)
         .foregroundStyle(Color.Artemis.secondaryLabel)
     }
-}
-
-#Preview {
-    CourseGridView(viewModel: DashboardViewModel(courseService: CourseServiceStub()))
 }
